fix UI issues
