--- libs/models/src/lib/models.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common'
-
-@Module({
-  controllers: [],
-  providers: [],
-  exports: [],
-})
-export class ModelsModule {}
