--- libs/models/src/lib/models.spec.ts
@@ -1,7 +1,0 @@
-import { models } from './models';
-
-describe('models', () => {
-  it('should work', () => {
-    expect(models()).toEqual('models');
-  });
-});
