--- libs/models/src/lib/models.ts
@@ -1,3 +1,0 @@
-export function models(): string {
-  return 'models';
-}
